First pass at extracting common compiler functionality into asttools
