+++ infra/outlines.py
@@ -1,12 +1,10 @@
 #!/usr/bin/env python3
 
 import argparse
-import ast
+import ast, asttools
 from dataclasses import dataclass
 import sys
 from typing import List
-
-groups = [ "Node", "Layout", "Draw" ]
 
 class Item: pass
 
@@ -93,73 +91,6 @@ def write_html(objs, indent=0):
             write_html(subs, indent=indent+4)
         print("</code>")
 
-def is_doc_string(cmd):
-    return isinstance(cmd, ast.Expr) and \
-        isinstance(cmd.value, ast.Constant) and isinstance(cmd.value.value, str)
-
-def is_sys_modules_hack(cmd):
-    return isinstance(cmd, ast.Assign) and len(cmd.targets) == 1 and \
-        isinstance(cmd.targets[0], ast.Attribute) and \
-        isinstance(cmd.targets[0].value, ast.Subscript) and \
-        isinstance(cmd.targets[0].value.value, ast.Attribute) and \
-        isinstance(cmd.targets[0].value.value.value, ast.Name) and \
-        cmd.targets[0].value.value.value.id == 'sys' and \
-        cmd.targets[0].value.value.attr == 'modules'
-
-def is_if_main(cmd):
-    return isinstance(cmd, ast.If) and isinstance(cmd.test, ast.Compare) and \
-        isinstance(cmd.test.left, ast.Name) and cmd.test.left.id == "__name__" and \
-        len(cmd.test.comparators) == 1 and isinstance(cmd.test.comparators[0], ast.Constant) and \
-        cmd.test.comparators[0].value == "__main__" and len(cmd.test.ops) == 1 and \
-        isinstance(cmd.test.ops[0], ast.Eq)
-
-def get_name(module, name):
-    assert isinstance(module, ast.Module)
-    for item in module.body:
-        if isinstance(item, ast.Import): pass
-        elif isinstance(item, ast.ImportFrom): pass
-        elif is_doc_string(item): pass
-        elif is_sys_modules_hack(item): pass
-        elif is_if_main(item): pass
-        elif isinstance(item, ast.ClassDef):
-            if item.name == name:
-                return item
-        elif isinstance(item, ast.FunctionDef):
-            if item.name == name:
-                return item
-        elif isinstance(item, ast.Assign) and len(item.targets) == 1:
-            if isinstance(item.targets[0], ast.Name):
-                if item.targets[0].id == name:
-                    return item
-            elif isinstance(item.targets[0], ast.Tuple):
-                assert isinstance(item.value, ast.Tuple)
-                for var, val in zip(item.targets[0].elts, item.value.elts):
-                    if var.id == name:
-                        return ast.Assign([var], val)
-            else:
-                raise Exception(ast.dump(cmd))
-            
-class ResolveImports(ast.NodeTransformer):
-    def visit_ImportFrom(self, cmd):
-        assert cmd.level == 0
-        assert cmd.module
-        assert all(name.asname is None for name in cmd.names)
-        names = [name.name for name in cmd.names]
-        filename = "src/{}.py".format(cmd.module)
-        objs = []
-
-        with open(filename) as file:
-            subast = ast.parse(file.read(), filename)
-
-        for name in names:
-            defn = get_name(subast, name)
-            if defn:
-                objs.append(defn)
-            else:
-                Exception("Could not find {} in module {}".format(name, cmd.module))
-
-        return objs
-
 def to_item(cmd):
     if is_sys_modules_hack(cmd): return
     elif is_if_main(cmd): return IfMain()
@@ -185,8 +116,7 @@ def to_item(cmd):
 
 def outline(tree):
     objs = []
-    assert isinstance(tree, ast.Module)
-    for cmd in tree.body:
+    for name, item in asttools.iter_defs(tree):
         item = to_item(cmd)
         if item: objs.append(item)
     return objs
@@ -198,9 +128,8 @@ if __name__ == "__main__":
     parser.add_argument("--html", action="store_true", default=False)
     args = parser.parse_args()
 
-    tree = ast.parse(args.file.read(), args.file.name)
-    tree2 = ResolveImports().visit(tree)
-    ol = outline(tree2)
+    tree = asttools.inline(asttools.parse(args.file.read(), args.file.name))
+    ol = outline(tree)
     if args.html:
         write_html(ol)
     else:
